requirements.txt
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
pyrootutils
pre-commit
rich
pytest
mmsegmentation>=1.0.0
